+++ telegram/__init__.py
@@ -1,3 +1,3 @@
-__version__ = '0.12.0'
+__version__ = '0.13.0'
 
 VERSION = __version__