+++ docs/source/changelog.rst
@@ -4,7 +4,7 @@ Changelog
 
 [unreleased]
 
-- Non-blocking login, see ``examples/get_me_non_blocking_login.py``.
+- Non-blocking login, see ``examples/get_me_non_blocking_login.py`` (thanks to @melfnt).
 
 [0.12.0] - 2020-03-29