+++ docs/source/changelog.rst
@@ -2,7 +2,7 @@
 Changelog
 =========
 
-[unreleased]
+[0.13.0] - 2020-11-16
 
 - Non-blocking login, see ``examples/get_me_non_blocking_login.py`` (thanks to @melfnt).
 - Better stop. python-teleram calls `close` and waits until tdlib is closed.